% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_cv.R
\name{cal_cv}
\alias{cal_cv}
\title{Calculate Coefficient of Variation}
\usage{
cal_cv(dnb, adjust_size = T)
}
\arguments{
\item{dnb}{a DNB object}

\item{adjust_size}{adjust sample size}
}
\value{
a DNB object
}
\description{
Calculate coefficient of variation among genes by sample groups
}
