% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_final.R
\name{cal_final}
\alias{cal_final}
\title{Calculate Final DNB Genes}
\usage{
cal_final(dnb, with_ctrl = T)
}
\arguments{
\item{dnb}{a DNB object}

\item{with_ctrl}{if consider control group. Default T.}
}
\value{
a DNB object
}
\description{
Compare and get the time point with highest score, and use that set of
DNB genes to recalculate DNB attibutes for all time points.
}
