% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_time_point_stats.R
\name{cal_time_point_stats}
\alias{cal_time_point_stats}
\title{Calculate time point statictics}
\usage{
cal_time_point_stats(dnb, name, FUN)
}
\arguments{
\item{dnb}{a DNB object}

\item{name}{statistical item name}

\item{FUN}{statistical function}

\item{...}{additional parameters to FUN}
}
\value{
a DNB object
}
\description{
Calculate time point statictics i.e. PCC, CV
}
