% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_correlation.R
\name{get_correlation}
\alias{get_correlation}
\title{Get correlation of a specific time point}
\usage{
get_correlation(dnb, time_point, cor_name = "correlation")
}
\arguments{
\item{dnb}{a DNB object}

\item{time_point}{character name of a time point}

\item{cor_name}{"correlation" or "correlation_ctrl"}
}
\value{
a correlation matrix
}
\description{
Get correlation of a specific time point
}
\examples{
\dontrun{
r <- get_correlation(dnb, "2", "correlation")
r_ctrl <- get_correlation(dnb, "2", "correlation_ctrl")
}
}
