% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DNB_attr.R
\name{get_DNB_attr}
\alias{get_DNB_attr}
\title{Get DNB Attributions}
\usage{
get_DNB_attr(dnb, time_point, genes_in, with_ctrl = T)
}
\description{
Calculate DNB score for given genes and return related parameters
}
