% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{new}
\alias{new}
\title{New DNB Object}
\usage{
new(data, time, group = NULL)
}
\arguments{
\item{data}{a gene expression matrix/data frame with sample as column names
and gene as row names.}

\item{time}{an factor of time points.}

\item{group}{a vector of 0 and 1 for control and case groups, Default NULL.}
}
\value{
a DNB object
}
\description{
Create a DNB object by specifying parameters
}
