% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_candidates.R
\name{search_candidates}
\alias{search_candidates}
\title{Search Candidate DNB genes}
\usage{
search_candidates(dnb, min_size = 2, max_size = Inf,
  included_genes = NULL, all = F, with_ctrl = T, verbose = T)
}
\arguments{
\item{dnb}{a DNB object}

\item{min_size}{minimum gene number in subtree. Default 2.}

\item{max_size}{maximum gene number in subtree. Default Inf.}

\item{included_genes}{genes to be included in subtree. Default NULL.}

\item{all}{should all \@code{included_genes} be in subtree. Default False.}

\item{with_ctrl}{if consider control group. Default T.}

\item{verbose}{output progress.}
}
\value{
a DNB object
}
\description{
Cluster by correlation, caculate DNB scores of all subtrees, and get best DNB
score for each time point
}
