% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_cor.R
\name{cal_cor}
\alias{cal_cor}
\title{Calculate Correlation}
\usage{
cal_cor(dnb, adjust_size = T)
}
\arguments{
\item{dnb}{a DNB object}

\item{adjust_size}{adjust sample size}
}
\value{
a DNB object
}
\description{
Calculate correlation among genes by sample groups
}
