% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DNB.R
\name{plot_DNB}
\alias{plot_DNB}
\title{Plot DNB}
\usage{
plot_DNB(dnb_df, file = "DNB_plot.pdf", height = 7, width = 7)
}
\arguments{
\item{dnb_df}{a data frame of DNB result.}

\item{file}{a character string of the output PDF file.}

\item{height, width}{the width and height of the graphics region in inches.}
}
\description{
Plot DNB attibutes from data frame
}
